models/clmr/requirements.txt
# pip install torch==1.12.1+cu116 torchvision==0.13.1+cu116 torchaudio==0.12.1 --extra-index-url https://download.pytorch.org/whl/cu116
pytorch-lightning==1.9.0
torchaudio-augmentations==0.2.4
simclr==1.0.2
torch-pitch-shift==1.2.4
soundfile==0.12.1
matplotlib==3.7.1
scikit-learn==1.2.2
tensorboard